requirements.txt
Django==4.2.7
tensorflow==2.13.0
Pillow==10.0.1
numpy==1.24.3
googletrans==4.0.0rc1
django-cors-headers==4.3.1